Refactored locations of Applications
