+++ src/module/applications/_module.mjs
@@ -1,9 +1,7 @@
-export * from "./actor-sheet/_module.mjs";
-export * from "./core/_module.mjs";
+export * as sidebar from "./sidebar/_module.mjs";
+export * as apps from "./apps/_module.mjs";
 export * as hooks from "./hooks/_module.mjs";
-export {DrawSteelItemSheet} from "./item-sheet.mjs";
-export {PowerRollDialog} from "./power-roll-dialog.mjs";
-export {TargetedConditionPrompt} from "./targeted-condition-prompt.mjs";
+export * as sheets from "./sheets/_module.mjs";
 
 import * as elements from "./elements/_module.mjs";
