+++ src/module/applications/_types.d.ts
@@ -1,0 +1,2 @@
+import "./apps/_types";
+import "./sheets/_types";
