+++ src/module/applications/apps/_module.mjs
@@ -1,0 +1,2 @@
+export {default as PowerRollDialog} from "./power-roll-dialog.mjs";
+export {default as TargetedConditionPrompt} from "./targeted-condition-prompt.mjs";
