+++ src/module/applications/elements/adopted-stylesheet-mixin.mjs
