+++ src/module/applications/elements/checkbox.mjs
