+++ src/module/applications/elements/slide-toggle.mjs
