+++ src/module/applications/hooks/combatantConfig.mjs
