+++ src/module/applications/hooks/tokenApplication.mjs
