+++ src/module/applications/sheets/npc.mjs
@@ -1,5 +1,5 @@
 import {systemID, systemPath} from "../../constants.mjs";
-import DrawSteelActorSheet from "./base.mjs";
+import DrawSteelActorSheet from "./actor-sheet.mjs";
 /** @import {FormSelectOption} from "../../../../foundry/client-esm/applications/forms/fields.mjs" */
 /** @import {DrawSteelActor} from "../../documents/actor.mjs"; */
