+++ src/module/applications/sidebar/_module.mjs
@@ -1,0 +1,1 @@
+export {default as DrawSteelCombatTracker} from "./combat-tracker.mjs";
