+++ src/module/applications/sidebar/combat-tracker.mjs
