--- src/module/apps/actor-sheet/_module.mjs
@@ -1,3 +1,0 @@
-export {default as DrawSteelActorSheet} from "./base.mjs";
-export {default as DrawSteelCharacterSheet} from "./character.mjs";
-export {default as DrawSteelNPCSheet} from "./npc.mjs";
