--- src/module/apps/core/_module.mjs
@@ -1,2 +1,0 @@
-export {default as DrawSteelActiveEffectConfig} from "./active-effect-config.mjs";
-export {default as DrawSteelCombatTracker} from "./combat-tracker.mjs";
