+++ src/module/rolls/power.mjs
@@ -1,4 +1,4 @@
-import {PowerRollDialog} from "../apps/power-roll-dialog.mjs";
+import PowerRollDialog from "../applications/apps/power-roll-dialog.mjs";
 import {systemPath} from "../constants.mjs";
 import {DrawSteelChatMessage} from "../documents/chat-message.mjs";
 import {DSRoll} from "./base.mjs";
